apps/api/pyproject.toml
[project]
name = "mergenix-api"
version = "0.1.0"
description = "Mergenix Genetic Analysis Platform — FastAPI Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pyotp>=2.9.0",
    "stripe>=11.3.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "structlog>=24.4.0",
    "sentry-sdk[fastapi]>=2.19.0",
    "slowapi>=0.1.9",
    "resend>=2.5.0",
    "qrcode[pil]>=8.0",
    "itsdangerous>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "aiosqlite>=0.20.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["app", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "S",    # flake8-bandit (security)
    "ASYNC",# flake8-async
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "S101",   # assert usage (fine in tests)
    "S105",   # hardcoded password (false positives in test data)
    "S106",   # hardcoded password (false positives in test data)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
